swars: Fixed invalid point index in build_glare()

+++ src/engindrwlstm.h
@@ -46,7 +46,12 @@ void draw_e_graphic_scale(int x, int y, int z, ushort frame, int radius, int int
 void draw_pers_e_graphic(struct Thing *p_thing, int x, int y, int z, int frame, int radius, int intensity);
 void FIRE_draw_fire(struct SimpleThing *p_sthing);
 void draw_bang(struct SimpleThing *p_pow);
-short draw_rot_object(int offset_x, int offset_y, int offset_z, struct SingleObject *point_object, struct Thing *p_thing);
+
+/** Adds primitives for a rotated 3D object to drawlist.
+ *
+ * @return Gives max depth value shifted to bucket index.
+ */
+int draw_rot_object(int offset_x, int offset_y, int offset_z, struct SingleObject *point_object, struct Thing *p_thing);
 short draw_rot_object2(int offset_x, int offset_y, int offset_z, struct SingleObject *point_object, struct Thing *p_thing);
 short draw_object(int x, int y, int z, struct SingleObject *point_object);
 void draw_vehicle_health(struct Thing *p_thing);
